+++ services/learn-card-network/brain-service/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @learncard/network-brain-service
 
+## 2.1.2
+
+### Patch Changes
+
+-   [#499](https://github.com/learningeconomy/LearnCard/pull/499) [`b6dd8b6`](https://github.com/learningeconomy/LearnCard/commit/b6dd8b65e295e0fe5122658cd351e1561d541523) Thanks [@cboydstun](https://github.com/cboydstun)! - count boosts route
+
 ## 2.1.1
 
 ### Patch Changes